requirements.txt
python-dotenv
discord.py
feedparser
databases[sqlite]
fuzzy-match